requirements.txt
httpx~=0.27.0
schema~=0.7.7
pillow~=10.4.0